requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.8.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
